dbgapi/msvc: Fix << shift
Fixes:

 [23/25] Building CXX object CMakeFiles\amd-dbgapi.dir\src\os_driver_kmd.cpp.obj
 D:\therock\src\rocm-systems\projects\rocdbgapi\src\os_driver_kmd.cpp(1370): warning C4334: '<<': result of 32-bit shift implicitly converted to 64 bits (was 64-bit shift intended?)
 [24/25] Linking CXX shared library rocm-dbgapi.dll

commit-id:3f69d182

+++ projects/rocdbgapi/src/os_driver_kmd.cpp
@@ -1368,7 +1368,7 @@ kmd_driver_t::agent_snapshot (os_agent_info_t *snapshots,
       agent.address_watch_supported
         = kmd_snap.capability & KMD_DBGR_CAP_WATCH_POINTS_SUPPORTED;
       agent.address_watch_register_count
-        = 1 << ((kmd_snap.capability
+        = static_cast<size_t>(1) << ((kmd_snap.capability
                  & KMD_DBGR_CAP_WATCH_POINTS_TOTALBITS_MASK)
                 >> KMD_DBGR_CAP_WATCH_POINTS_TOTALBITS_SHIFT);
       agent.precise_memory_supported
